Removed old GT_Cover_ControlsWork
